backend/requirements.txt

fastapi
uvicorn
pandas
faiss-cpu
numpy
openai>=1.12.0
python-dotenv
requests
langchain_openai